Add LineEditCompleter to Tab_Game
